requirements.txt
click==8.1.3
Flask==2.2.5
imageio==2.28.1
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
networkx==2.6.3
numpy==1.21.6
packaging==23.1
Pillow==9.5.0
PyWavelets==1.3.0
scikit-image==0.19.3
scipy==1.7.3
tifffile==2021.11.2
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
gunicorn==20.1.0
tensorflow==2.16.1
